refactor: delete dead command classes, util/, and unused helpers (-1000 lines)
Strip api/ to minimal stubs (data types only), delete all 30+ command
classes that are fully replaced by _vizio.py direct HTTP calls. Delete
util/ package (duplicated in apps.py). Remove get_value_from_path from
helpers.py. Update cli.py imports to canonical locations.

Deleted: api/base.py, api/remote.py, api/settings.py, util/
Stripped: api/_protocol.py, api/item.py, api/input.py, api/pair.py, api/apps.py
