--- pyvizio/api/base.py
@@ -1,90 +1,0 @@
-"""Vizio SmartCast API base commands."""
-
-from __future__ import annotations
-
-from typing import Any
-
-
-class CommandBase:
-    """Base command to send data to Vizio device."""
-
-    def __init__(self, url: str = "") -> None:
-        """Initialize base command to send data to Vizio device."""
-        self._url = url
-
-    def __repr__(self) -> str:
-        return f"{type(self).__name__}({self.__dict__})"
-
-    def __eq__(self, other) -> bool:
-        return self is other or self.__dict__ == other.__dict__
-
-    def to_dict(self) -> dict[str, Any]:
-        """Return public attributes as dict for JSON serialization."""
-
-        def _serialize(obj: Any) -> Any:
-            if isinstance(obj, list):
-                return [_serialize(item) for item in obj]
-            if hasattr(obj, "__dict__") and not isinstance(obj, type):
-                return {
-                    k: _serialize(v)
-                    for k, v in obj.__dict__.items()
-                    if not k.startswith("_")
-                }
-            return obj
-
-        return {
-            k: _serialize(v) for k, v in self.__dict__.items() if not k.startswith("_")
-        }
-
-    @property
-    def _method(self) -> str:
-        """Get command method."""
-        return "put"
-
-    @property
-    def url(self) -> str:
-        """Get endpoint for command."""
-        return self._url
-
-    @url.setter
-    def url(self, new_url: str) -> None:
-        """Set endpoint for command."""
-        self._url = new_url
-
-    def get_url(self) -> str:
-        """Get endpoint for command."""
-        return self._url
-
-    def get_method(self) -> str:
-        return self._method
-
-    def process_response(self, json_obj: dict[str, Any]) -> Any:
-        """Always return True when there is no custom process_response method for subclass."""
-        return True
-
-
-class InfoCommandBase(CommandBase):
-    """Base command to get data from Vizio device."""
-
-    def __init__(self, url: str = "") -> None:
-        """Initialize base command to get data from Vizio device."""
-        super().__init__(url)
-
-    @property
-    def _method(self) -> str:
-        """Get command method."""
-        return "get"
-
-    @property
-    def url(self) -> str:
-        """Get endpoint for command."""
-        return self._url
-
-    @url.setter
-    def url(self, new_url: str) -> None:
-        """Set endpoint for command."""
-        self._url = new_url
-
-    def process_response(self, json_obj: dict[str, Any]) -> Any:
-        """Always return None when there is no custom process_response method for subclass."""
-        return None
