--- pyvizio/api/remote.py
@@ -1,44 +1,0 @@
-"""Vizio SmartCast API command and class for emulating remote key presses."""
-
-from __future__ import annotations
-
-from typing import Any
-
-from pyvizio.api._protocol import ENDPOINT, KEY_ACTION
-from pyvizio.api.base import CommandBase
-
-
-class KeyPressEvent:
-    """Emulated remote key press."""
-
-    def __init__(
-        self, key_code: tuple[int, int], action: str = KEY_ACTION["PRESS"]
-    ) -> None:
-        """Initialize emulated remote key press."""
-        self.CODESET: int = key_code[0]
-        self.CODE: int = key_code[1]
-        self.ACTION: str = action
-
-    def __repr__(self) -> str:
-        return f"{type(self).__name__}({self.__dict__})"
-
-    def __eq__(self, other) -> bool:
-        return self is other or self.__dict__ == other.__dict__
-
-
-class EmulateRemoteCommand(CommandBase):
-    """Command to emulate remote key press."""
-
-    def __init__(self, key_codes: list[tuple[int, int]], device_type: str) -> None:
-        """Initialize command to emulate remote key press."""
-        super().__init__(ENDPOINT[device_type]["KEY_PRESS"])
-
-        # noinspection SpellCheckingInspection
-        self.KEYLIST: list[KeyPressEvent] = []
-
-        for key_code in key_codes:
-            self.KEYLIST.append(KeyPressEvent(key_code))
-
-    def process_response(self, json_obj: dict[str, Any]) -> bool:
-        """Return True on successful key press."""
-        return True
