--- pyvizio/api/settings.py
@@ -1,176 +1,0 @@
-"""Vizio SmartCast API commands for audio settings."""
-
-from __future__ import annotations
-
-from typing import Any
-
-from pyvizio.api._protocol import (
-    ENDPOINT,
-    TYPE_LIST,
-    TYPE_MENU,
-    TYPE_SLIDER,
-    TYPE_VALUE,
-    TYPE_X_LIST,
-    ResponseKey,
-)
-from pyvizio.api.item import Item, ItemCommandBase, ItemInfoCommandBase
-from pyvizio.helpers import dict_get_case_insensitive
-
-
-class GetAllSettingTypesCommand(ItemInfoCommandBase):
-    """Command to get list of all setting types."""
-
-    def __init__(self, device_type: str) -> None:
-        """Initialize command to get list of all setting types."""
-        super().__init__(device_type, "SETTINGS")
-        self._url = f"{ENDPOINT[device_type]['SETTINGS']}"
-
-    def process_response(self, json_obj: dict[str, Any]) -> list[str]:
-        """Return response to command to get list of all setting types."""
-        items = [
-            Item(item)
-            for item in dict_get_case_insensitive(json_obj, ResponseKey.ITEMS, [])
-        ]
-        return [
-            item.c_name
-            for item in items
-            if item.type.lower() == TYPE_MENU
-            and item.c_name not in ("cast", "input", "devices", "network")
-        ]
-
-
-class GetAllSettingsCommand(ItemInfoCommandBase):
-    """Command to get list of all setting names and corresponding values."""
-
-    def __init__(self, device_type: str, setting_type: str) -> None:
-        """Initialize command to get list of all setting names and corresponding values."""
-        super().__init__(device_type, "SETTINGS")
-        self.setting_type = setting_type.lower()
-        self._url = f"{ENDPOINT[device_type]['SETTINGS']}/{setting_type}"
-
-    def process_response(self, json_obj: dict[str, Any]) -> dict[str, int | str]:
-        """Return response to command to get list of all setting names and corresponding values."""
-        items = [
-            Item(item)
-            for item in dict_get_case_insensitive(json_obj, ResponseKey.ITEMS, [])
-        ]
-        return {
-            item.c_name: item.value
-            for item in items
-            if item.type.lower() in (TYPE_LIST, TYPE_SLIDER, TYPE_VALUE)
-        }
-
-
-class GetSettingCommand(ItemInfoCommandBase):
-    """Command to get value of a setting by name."""
-
-    def __init__(self, device_type: str, setting_type: str, setting_name: str) -> None:
-        """Initialize command to get value of a setting by name."""
-        super().__init__(device_type, "SETTINGS", 0)
-        self.item_name = setting_name.lower()
-        self.setting_type = setting_type.lower()
-        self._url = f"{ENDPOINT[device_type]['SETTINGS']}/{setting_type}/{setting_name}"
-
-
-class GetAllSettingsOptionsCommand(ItemInfoCommandBase):
-    """Command to get list of all setting names and corresponding options."""
-
-    def __init__(self, device_type: str, setting_type: str) -> None:
-        """Initialize command to get list of all setting names and corresponding options."""
-        super().__init__(device_type, "SETTINGS_OPTIONS")
-        self.setting_type = setting_type.lower()
-        self._url = f"{ENDPOINT[device_type]['SETTINGS_OPTIONS']}/{setting_type}"
-
-    def process_response(
-        self, json_obj: dict[str, Any]
-    ) -> dict[str, list[str] | dict[str, int | None]]:
-        """Return response to command to get list of all setting names and corresponding options."""
-        items = [
-            Item(item)
-            for item in dict_get_case_insensitive(json_obj, ResponseKey.ITEMS, [])
-        ]
-
-        settings_options: dict[str, list[str] | dict[str, int | None]] = {}
-        for item in items:
-            if item.type.lower() in (TYPE_SLIDER, TYPE_VALUE):
-                d: dict[str, int | None] = {"min": item.min, "max": item.max}
-                if item.center is not None:
-                    d["default"] = item.center
-                settings_options[item.c_name] = d
-            elif item.type.lower() == TYPE_LIST:
-                settings_options[item.c_name] = item.choices.copy()
-
-        return settings_options
-
-
-class GetSettingOptionsCommand(GetAllSettingsOptionsCommand):
-    """Command to get options of a setting by name."""
-
-    def __init__(self, device_type: str, setting_type: str, setting_name: str) -> None:
-        """Initialize command to get options of a setting by name."""
-        self.setting_name = setting_name
-        self.setting_type = setting_type.lower()
-        super().__init__(device_type, self.setting_type)
-
-    def process_response(  # type: ignore[override]
-        self, json_obj: dict[str, Any]
-    ) -> list[str] | dict[str, int | None] | None:
-        """Return response to command to get options of a setting by name."""
-        return super().process_response(json_obj).get(self.setting_name)
-
-
-class GetAllSettingsOptionsXListCommand(ItemInfoCommandBase):
-    """Command to get list of all setting names and corresponding options for settings of type XList."""
-
-    def __init__(self, device_type: str, setting_type: str) -> None:
-        """Initialize command to get list of all setting names and corresponding options for settings of type XList."""
-        super().__init__(device_type, "SETTINGS")
-        self.setting_type = setting_type.lower()
-        self._url = f"{ENDPOINT[device_type]['SETTINGS']}/{setting_type}"
-
-    def process_response(self, json_obj: dict[str, Any]) -> dict[str, list[str]]:
-        """Return response to command to get list of all setting names and corresponding options for settings of type XList."""
-        items = [
-            Item(item)
-            for item in dict_get_case_insensitive(json_obj, ResponseKey.ITEMS, [])
-        ]
-
-        return {
-            item.c_name: item.choices
-            for item in items
-            if item.type and item.type.lower() == TYPE_X_LIST
-        }
-
-
-class GetSettingOptionsXListCommand(GetAllSettingsOptionsXListCommand):
-    """Command to get options of an audio setting by name (used for setting of type XList)."""
-
-    def __init__(self, device_type: str, setting_type: str, setting_name: str) -> None:
-        """Initialize command to get options of an audio setting by name (used for setting of type XList)."""
-        self.setting_name = setting_name
-        self.setting_type = setting_type.lower()
-        super().__init__(device_type, self.setting_type)
-
-    def process_response(self, json_obj: dict[str, Any]) -> list[str] | None:  # type: ignore[override]
-        """Return response to command to get options of an audio setting by name (used for setting of type XList)."""
-        return super().process_response(json_obj).get(self.setting_name)
-
-
-class ChangeSettingCommand(ItemCommandBase):
-    """Command to set value of a setting by name to new value."""
-
-    def __init__(
-        self,
-        device_type: str,
-        id: int,
-        setting_type: str,
-        setting_name: str,
-        value: int | str,
-    ) -> None:
-        """Initialize command to set value of a setting by name to new value."""
-        super().__init__(device_type, "SETTINGS", id, value)
-        self._url = f"{ENDPOINT[device_type]['SETTINGS']}/{setting_type}/{setting_name}"
-
-    def process_response(self, json_obj: dict[str, Any]) -> bool:
-        """Return True on successful setting change."""
-        return True
