+++ tests/test_helpers.py
@@ -5,7 +5,6 @@ import pytest
 from pyvizio.helpers import (
     async_to_sync,
     dict_get_case_insensitive,
-    get_value_from_path,
 )
 
 
@@ -27,44 +26,6 @@ class TestDictGetCaseInsensitive:
     def test_missing_key_returns_default(self, default):
         result = dict_get_case_insensitive({"key": "value"}, "other", default)
         assert result is default
-
-
-class TestGetValueFromPath:
-    def test_single_level_path(self):
-        assert (
-            get_value_from_path({"model_name": "V505-G9"}, [["model_name"]])
-            == "V505-G9"
-        )
-
-    def test_single_level_case_insensitive(self):
-        assert (
-            get_value_from_path({"MODEL_NAME": "V505-G9"}, [["model_name"]])
-            == "V505-G9"
-        )
-
-    def test_missing_path_returns_none(self):
-        assert get_value_from_path({"other": "value"}, [["model_name"]]) is None
-
-    def test_multiple_paths_first_match(self):
-        data = {"model_name": "V505-G9"}
-        paths = [["model_name"], ["model"]]
-        assert get_value_from_path(data, paths) == "V505-G9"
-
-    def test_multiple_paths_fallback(self):
-        """Second path matches when first doesn't."""
-        data = {"model": "V505-G9"}
-        paths = [["model_name"], ["model"]]
-        assert get_value_from_path(data, paths) == "V505-G9"
-
-    @pytest.mark.parametrize(
-        "data,paths",
-        [
-            ({}, [["key"]]),
-            ({"key": "val"}, []),
-        ],
-    )
-    def test_returns_none(self, data, paths):
-        assert get_value_from_path(data, paths) is None
 
 
 class TestAsyncToSync:
